implement WebSocket API
following the esp_websocket_client found here[^1][^2], this commit
wraps the mentioned C library in Rust to expose a safe WebSocket
interface.
The exposed API copies the previously implemented MQTT API. On creating
a `EspWebSocketClientConfig` and passing it to a `EspWebSocketClient`,
the user can call `send` to send data. Manual sending of fragmented data,
Ping/Pong, or Connect/Disconnect are not supported due to the underlying
C library not supporting this, either.
Receiving is done in a new thread with the returned
`EspWebSocketConnection` and its blocking `next` function call, which
wakes on receiving one of the events defined in `WebSocketEventType`.
The `EspWebSocketClient` takes a `time::Duration` timeout since the
underlying C lib requires this but the defined interface in the
`embedded_svc` create does not provide it as a parameter.
On dropping the `EspWebSocketClient` the C functions
`esp_websocket_client_close` and `esp_websocket_client_destroy` are
being called, cleanly closing the connection to the WebSocket server.

[^1]: https://espressif.github.io/esp-protocols/esp_websocket_client/
[^2]: https://github.com/espressif/esp-protocols/