Websocket implementation (#69)
* implement WebSocket API

following the esp_websocket_client found here[^1][^2], this commit
wraps the mentioned C library in Rust to expose a safe WebSocket
interface.
The exposed API copies the previously implemented MQTT API. On creating
a `EspWebSocketClientConfig` and passing it to a `EspWebSocketClient`,
the user can call `send` to send data. Manual sending of fragmented data,
Ping/Pong, or Connect/Disconnect are not supported due to the underlying
C library not supporting this, either.
Receiving is done in a new thread with the returned
`EspWebSocketConnection` and its blocking `next` function call, which
wakes on receiving one of the events defined in `WebSocketEventType`.
The `EspWebSocketClient` takes a `time::Duration` timeout since the
underlying C lib requires this but the defined interface in the
`embedded_svc` create does not provide it as a parameter.
On dropping the `EspWebSocketClient` the C functions
`esp_websocket_client_close` and `esp_websocket_client_destroy` are
being called, cleanly closing the connection to the WebSocket server.

[^1]: https://espressif.github.io/esp-protocols/esp_websocket_client/
[^2]: https://github.com/espressif/esp-protocols/

* Add missing fields in `EspWebSocketClientConfig`

Missing fields for the used interface and SSL data have been added.
`From` on `(esp_websocket_client_config_t, RawCstrs)` had to be
converted to a `TryFrom` as the used interface field should be a 6 ASCII
character long array, but cannot be validated anywhere else.
Some returned Errors have been converted to `anyhow::Error` types for
simplicity.

* Fix bug receiving message assumed to end in '\0'

* annotate unused field, remove unused import

annotate the unused callback field in `EspWebSocketClient` to allow dead
code. it exists so that the callback is not dropped, as it is used
outside of the scope it is instantiated in (in the c lib).
remove an unused import

* temporarily enable experimental feature for embedded-svc

* adjust for default ESP-IDF version being 4.3.2

ESP-IDF v4.3.2 has a few differences to v4.4, which are addressed here
by conditional compilation.

* cfg: remove 'all' for websocket

* derive usual traits on EspWebSocketTransport

* use EspError instead of anyhow

* restructure EspWebSocketClientConfig

take a mandatory uri on creating a new EspWebSocketClient. this makes
the fields uri, host, port, and path on the config unnecessary

* task_prio and task_stack are usize

* use only one condvar instead of two

Co-authored-by: Ivan Markov <ivan.markov@gmail.com>