Wifi: Split status tuple for AP and STA status into two separate fields
This is in preparation for async, which will want to wait for one of these to change
but not the other.

Also, I think it increases legibility and decreases risk of error due to confusing which
of .0 and .1 in the tuple is AP and STA statuses.

Note: WifiDriver::is_connected() returns false if the AP is enabled but not started even if
the STA is connected. This seems odd to me, shouldn't is_connected() only be about if the
Wifi is connected as STA?