store RawCstrs in EspWebSocketClient
esp_websocket_client_config_t.cert_pem (and client_cert/client_key) are set to pointers to CStrings created in a RawCstrs instance that is dropped at the end of the EspWebSocketClientConfig to esp_websocket_client_config_t conversion function.

But the functions that use those pointers (esp_transport_ssl_set_cert_data_der(), etc.) all say, "Note that, this function stores the pointer to data, rather than making a copy. So this data must remain valid until after the connection is cleaned up."

So the RawCstrs needs to live as long as the underlying connection. This change ensures that by moving it into the EspWebSocketClient.