set esp_websocket_client_config_t.cert_len to `0` for PEM-encoded certificate
When configuring esp_websocket_client_config_t with a PEM-encoded certificate, esp_websocket_client explains cert_pem to be set to the nul-terminated certificate string and cert_len to be set to `0`.

EspWebsocketClient currently sets cert_pem correctly but sets cert_len to the length of the string.  And that causes esp_websocket_client to treat cert_pem as being DER-encoded rather than PEM-encoded, because despite the name, cert_pem can contain either a PEM-encoded or a DER-encoded certificate, and esp_websocket_client uses cert_len to determine which it is: if cert_len is `0`, then cert_pem is treated as PEM-encoded; otherwise, it's treated as DER-encoded.

More specifically, setting esp_websocket_client_config_t.cert_len for a PEM-encoded certificate causes esp_websocket_client_create_transport() to call esp_transport_ssl_set_cert_data_der(), which sets esp_tls_cfg.cacert_bytes to the length not including the trailing nul, although the docs say it should be "including NULL-terminator in case of PEM format."

And mbedtls_x509_crt_parse() then doesn't find the trailing nul and tries to parse the certificate as DER, which fails:

> E (11701) esp-tls-mbedtls: mbedtls_x509_crt_parse of CA cert returned -0x2180
> E (11701) esp-tls-mbedtls: Failed to set client configurations, returned [0x8015] (ESP_ERR_MBEDTLS_X509_CRT_PARSE_FAILED)
> E (11711) esp-tls: create_ssl_handle failed
> E (11711) esp-tls: Failed to open new connection
> E (11721) transport_base: Failed to open a new connection
> E (11731) transport_ws: Error connecting to host example.websocket.org:443
> E (11741) websocket_client: esp_transport_connect() failed with -1, transport_error=ESP_ERR_MBEDTLS_X509_CRT_PARSE_FAILED, tls_error_code=8576, tls_flags=0, esp_ws_handshake_status_code=0, errno=119

Rather than adding a cert_len field to EspWebSocketClientConfig and replicating the non-obvious interface of esp_websocket_client_config_t, this commit adds an EspWebSocketX509Encoding enum with PEM and DER variants, renames EspWebSocketClientConfig.cert_pem to EspWebSocketClientConfig.server_cert, and makes it an EspWebSocketX509Encoding, so the API consumer must specify explicitly the encoding for the server certificate.

Internally, `impl<'a> TryFrom<&'a EspWebSocketClientConfig<'a>> for (esp_websocket_client_config_t, RawCstrs)` then converts the value into the appropriate values for esp_websocket_client_config_t.cert_pem and cert_len.

Based on my reading of the source, the same thing needs to be done for the client_cert and client_key fields, so I've done the same for them. I haven't yet tested those fields, nor have I tested a DER-encoded cert, but I can confirm that this change fixes the use of PEM-encoded certs.