disable component manager when building with ESP-IDF 4.*
Disable the component manager when building with ESP-IDF 4.* and add the esp_websocket_client extra component by default to fix building the websocket_client example in CI for both 4.* and 5.*.

The problem with this solution is that it breaks building examples with ESP-IDF 4.* unless you disable the component manager. There should be a better way to do this, such as conditionally configuring esp-idf-sys to add the extra component only when building with ESP-IDF 5.*. But I haven't found a way to do that yet.