Make async mqtt client implement Send. (#461)
There were already some Send implementations.
However, they weren't completely correct, which made it impossible
to use the async mqtt client in e.g. a tokio worker thread.

Note that it could be used in the async main task, because that never
sends off any work to other worker threads, and thus Send is not required.
The errors would surface when trying to use tokio::spawn with an async mqtt client.