Added default option for httpd_config_t::task_caps (#489)
IDF v5.3.1 introduced a new configuration option for the HTTP server to
specify the FreeRTOS memory capabilities for the task's stack memory.
The official default value is `MALLOC_CAP_INTERNAL | MALLOC_CAP_8BIT`
(as per esp_http_server.h). This translates to `(1 << 11) | (1 << 2)`
which is used here.

Currently this value is not exposed via `http::server::Configuration`.